src/backend/offline/requirements.txt
requests
beautifulsoup4
lxml
trafilatura
tqdm
python-dotenv
chromadb
openai
tiktoken
lxml[html_clean]
